+++ triggers/grpc/src/Workflows.ts
@@ -1,0 +1,5 @@
+import type Workflows from "./types/Workflows";
+
+const workflows: Workflows = {};
+
+export default workflows;