+++ nodes/web/api-call@1.0.0/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @nanoservice-ts/api-call
 
+## 0.1.18
+
+### Patch Changes
+
+- Improved the NanoService base class to accept a InputType. This force developer to always create a type to define the Node handle input. Added unit test for pending projects like if-else and api-call.
+- Updated dependencies
+  - @nanoservice-ts/runner@0.1.15
+
 ## 0.1.17
 
 ### Patch Changes