+++ triggers/http/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @nanoservice-ts/trigger-http
 
+## 0.0.21
+
+### Patch Changes
+
+- Improved the NanoService base class to accept a InputType. This force developer to always create a type to define the Node handle input. Added unit test for pending projects like if-else and api-call.
+- Updated dependencies
+  - @nanoservice-ts/if-else@0.0.19
+  - @nanoservice-ts/api-call@0.1.18
+  - @nanoservice-ts/runner@0.1.15
+
 ## 0.0.20
 
 ### Patch Changes